refactor the fold structure
