+++ src/helper/array.ts
