+++ src/index.ts
@@ -1,7 +1,7 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { enableProdMode } from '@angular/core';
 import { environment } from './environments/environment';
-import { MemoryModule } from './game/memory.module';
+import { MemoryModule } from './memorygame/memory.module';
 
 if (environment.production) {
   enableProdMode();
