+++ src/memorygame/dashboard/info.component.css
