+++ src/memorygame/dashboard/info.component.html
