+++ src/memorygame/dashboard/info.component.ts
