+++ src/memorygame/memory.component.html
