+++ src/memorygame/memory.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 
-import { ReduxConfigModule } from './store';
+import { ReduxConfigModule } from '../store';
 
 import { MemoryComponent } from './memory.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
@@ -11,7 +11,7 @@ import { ChessboardComponent } from './chessboard/chessboard.component';
 import { CardComponent } from './chessboard/card.component';
 import { StatusComponent } from './status/status.component';
 
-import { GameActions } from './store/action';
+import { GameActions } from '../store/action';
 
 @NgModule({
     imports: [
