+++ src/memorygame/status/status.component.html
