+++ src/store/interface/ICard.ts
@@ -1,0 +1,6 @@
+export interface ICard {
+    _id?: number;
+    flipped: boolean;
+    name: string;
+    url?: string;
+}
