+++ src/store/interface/index.ts
@@ -1,0 +1,4 @@
+
+export * from './ICard';
+export * from './IState';
+export * from './status';
