+++ src/store/interface/status.ts
